Extracted createOneRecordMutation
