+++ packages/twenty-front/src/modules/object-record/hooks/useGenerateCreateOneRecordMutation.ts
@@ -5,12 +5,9 @@ import { EMPTY_MUTATION } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { objectMetadataItemsState } from '@/object-metadata/states/objectMetadataItemsState';
 import { ObjectMetadataItem } from '@/object-metadata/types/ObjectMetadataItem';
 import { mapObjectMetadataToGraphQLQuery } from '@/object-metadata/utils/mapObjectMetadataToGraphQLQuery';
+import { getCreateOneRecordMutationResponseField } from '@/object-record/utils/getCreateOneRecordMutationResponseField';
 import { isUndefinedOrNull } from '~/utils/isUndefinedOrNull';
 import { capitalize } from '~/utils/string/capitalize';
-
-export const getCreateOneRecordMutationResponseField = (
-  objectNameSingular: string,
-) => `create${capitalize(objectNameSingular)}`;
 
 export const useGenerateCreateOneRecordMutation = ({
   objectMetadataItem,
