+++ packages/twenty-front/src/modules/object-record/utils/getCreateOneRecordMutationResponseField.ts
@@ -1,0 +1,5 @@
+import { capitalize } from '~/utils/string/capitalize';
+
+export const getCreateOneRecordMutationResponseField = (
+  objectNameSingular: string,
+) => `create${capitalize(objectNameSingular)}`;
