+++ packages/twenty-front/src/modules/views/hooks/internal/usePersistViewFieldRecords.ts
@@ -7,17 +7,18 @@ import { triggerUpdateRecordOptimisticEffect } from '@/apollo/optimistic-effect/
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { useCreateOneRecordMutation } from '@/object-record/hooks/useCreateOneRecordMutation';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { GraphQLView } from '@/views/types/GraphQLView';
 import { ViewField } from '@/views/types/ViewField';
 
 export const usePersistViewFieldRecords = () => {
-  const {
-    updateOneRecordMutation,
-    createOneRecordMutation,
-    getRecordFromCache,
-    objectMetadataItem,
-  } = useObjectMetadataItem({
+  const { updateOneRecordMutation, getRecordFromCache, objectMetadataItem } =
+    useObjectMetadataItem({
+      objectNameSingular: CoreObjectNameSingular.ViewField,
+    });
+
+  const { createOneRecordMutation } = useCreateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewField,
   });
