+++ packages/twenty-front/src/modules/views/hooks/internal/usePersistViewFilterRecords.ts
@@ -7,6 +7,7 @@ import { triggerUpdateRecordOptimisticEffect } from '@/apollo/optimistic-effect/
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
 import { useObjectMetadataItems } from '@/object-metadata/hooks/useObjectMetadataItems';
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
+import { useCreateOneRecordMutation } from '@/object-record/hooks/useCreateOneRecordMutation';
 import { ObjectRecord } from '@/object-record/types/ObjectRecord';
 import { GraphQLView } from '@/views/types/GraphQLView';
 import { ViewFilter } from '@/views/types/ViewFilter';
@@ -14,11 +15,14 @@ import { ViewFilter } from '@/views/types/ViewFilter';
 export const usePersistViewFilterRecords = () => {
   const {
     updateOneRecordMutation,
-    createOneRecordMutation,
     deleteOneRecordMutation,
     objectMetadataItem,
     getRecordFromCache,
   } = useObjectMetadataItem({
+    objectNameSingular: CoreObjectNameSingular.ViewFilter,
+  });
+
+  const { createOneRecordMutation } = useCreateOneRecordMutation({
     objectNameSingular: CoreObjectNameSingular.ViewFilter,
   });
